improved size & doc report
